(REF) APIv4 FieldSpec - Extract ArrayFormatTrait. Add loadArray().
